+++ sdk/typescript/src/auth/private-key-auth.ts
@@ -1,7 +1,6 @@
 // Copyright (c) RoochNetwork
 // SPDX-License-Identifier: Apache-2.0
 
-import { Bytes } from '../types'
 import { Keypair } from '../utils/crypto'
 import { IAuthorization, IAuthorizer } from './interface'
 
@@ -14,7 +13,7 @@ export class PrivateKeyAuth implements IAuthorizer {
     this.pk = pk
   }
 
-  async auth(data: Bytes): Promise<IAuthorization> {
+  async auth(data: Uint8Array): Promise<IAuthorization> {
     const sign = await this.pk.signMessage(data)
 
     return {
