+++ sdk/typescript/src/provider/interface.ts
@@ -1,6 +1,6 @@
 // Copyright (c) RoochNetwork
 // SPDX-License-Identifier: Apache-2.0
-import { Arg, TypeTag, FunctionId, AnnotatedFunctionResultView, Bytes } from '../types'
+import { Arg, TypeTag, FunctionId, rpcTypes, Bytes } from '../types'
 
 export interface IProvider {
   getRpcApiVersion(): Promise<string | undefined>
@@ -11,7 +11,7 @@ export interface IProvider {
     funcId: FunctionId,
     tyArgs?: TypeTag[],
     args?: Arg[],
-  ): Promise<AnnotatedFunctionResultView>
+  ): Promise<rpcTypes.AnnotatedFunctionResultView>
 
   sendRawTransaction(playload: Bytes): Promise<string>
 }
