+++ sdk/typescript/src/provider/json-rpc-provider.ts
@@ -5,20 +5,21 @@ import fetch from 'isomorphic-fetch'
 import { HTTPTransport, RequestManager } from '@open-rpc/client-js'
 import { JsonRpcClient } from '../generated/client'
 import { Connection, LocalNetConnection } from './connection'
-import { bytes } from '../types/bcs'
 import {
   FunctionId,
   TypeTag,
   Arg,
-  AnnotatedFunctionResultView,
-  TransactionView,
-  AnnotatedStateView,
+  Bytes,
+  rpcTypes,
+  // AnnotatedFunctionResultView,
+  // TransactionView,
+  // AnnotatedStateView,
   // TransactionExecutionInfoView,
 } from '../types'
 import { functionIdToStirng, typeTagToString, encodeArg, toHexString } from '../utils'
 
 import { ROOCH_LOCAL_CHIAN_ID } from '../constants'
-import { TransactionResultPageView } from '../generated/client/types.ts'
+//import { TransactionResultPageView } from '../generated/client/types.ts'
 
 /**
  * Configuration options for the JsonRpcProvider. If the value of a field is not provided,
@@ -96,7 +97,7 @@ export class JsonRpcProvider {
     funcId: FunctionId,
     tyArgs?: TypeTag[],
     args?: Arg[],
-  ): Promise<AnnotatedFunctionResultView> {
+  ): Promise<rpcTypes.AnnotatedFunctionResultView> {
     const tyStrArgs = tyArgs?.map((v) => typeTagToString(v))
     const bcsArgs = args?.map((arg) => toHexString(encodeArg(arg))) as any
 
@@ -109,13 +110,13 @@ export class JsonRpcProvider {
 
   // Send the signed transaction in bcs hex format
   // This method does not block waiting for the transaction to be executed.
-  async sendRawTransaction(playload: bytes): Promise<string> {
+  async sendRawTransaction(playload: Bytes): Promise<string> {
     return this.client.rooch_sendRawTransaction(playload)
   }
 
-  async getTransactionsByHash(tx_hashes: string[]): Promise<TransactionView | null[]> {
-    return await this.client.rooch_getTransactionsByHash(tx_hashes)
-  }
+  //async getTransactionsByHash(tx_hashes: string[]): Promise<TransactionView | null[]> {
+  //  return await this.client.rooch_getTransactionsByHash(tx_hashes)
+  //}
 
   // async getTransactionInfosByHash(
   //   txHashes: string[],
@@ -124,13 +125,13 @@ export class JsonRpcProvider {
   // }
 
   // Get the annotated states by access_path The annotated states include the decoded move value of the state
-  async getAnnotatedStates(accessPath: string): Promise<AnnotatedStateView | null[]> {
+  async getAnnotatedStates(accessPath: string): Promise<rpcTypes.AnnotatedStateView | null[]> {
     return await this.client.rooch_getAnnotatedStates(accessPath)
   }
 
-  async getTransactionsByOrder(cursor: number, limit: number): Promise<TransactionResultPageView> {
-    return this.client.rooch_getTransactionsByOrder(cursor, limit)
-  }
+  //async getTransactionsByOrder(cursor: number, limit: number): Promise<rpcTypes.TransactionResultPageView> {
+  //  return this.client.rooch_getTransactionsByOrder(cursor, limit)
+  // }
 
   // TODO: wait bcs
