+++ sdk/typescript/src/types/bytes.ts
@@ -1,5 +1,4 @@
 // Copyright (c) RoochNetwork
 // SPDX-License-Identifier: Apache-2.0
-import { type bytes } from './bcs'
 
-export type Bytes = bytes
+export type Bytes = Uint8Array
