+++ sdk/typescript/src/types/index.ts
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 export * from './hex'
 export * from './rooch'
-export * from './rpc'
 export * from './bytes'
+
+export * as rpcTypes from './rpc'
 export * as bcsTypes from './bcs'
